# common-platform-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/organization'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/branches/{branchId}/departments/{departmentId}/positions/{positionId}':
    get:
      tags:
        - Branch Position Management
      summary: Get branch position by ID
      description: Returns a position of a specific branch department based on its ID
      operationId: getBranchPositionById
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: integer
            format: int64
        - name: departmentId
          in: path
          description: ID of the department
          required: true
          schema:
            type: integer
            format: int64
        - name: positionId
          in: path
          description: ID of the position to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '404':
          description: 'Branch, department, or position not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
    put:
      tags:
        - Branch Position Management
      summary: Update branch position
      description: >-
        Updates an existing position of a specific branch department with the
        provided details
      operationId: updateBranchPosition
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: integer
            format: int64
        - name: departmentId
          in: path
          description: ID of the department
          required: true
          schema:
            type: integer
            format: int64
        - name: positionId
          in: path
          description: ID of the position to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchPositionDTO'
        required: true
      responses:
        '200':
          description: Branch position successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '400':
          description: Invalid branch position data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '404':
          description: 'Branch, department, or position not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
    delete:
      tags:
        - Branch Position Management
      summary: Delete branch position
      description: Deletes a position of a specific branch department based on its ID
      operationId: deleteBranchPosition
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: integer
            format: int64
        - name: departmentId
          in: path
          description: ID of the department
          required: true
          schema:
            type: integer
            format: int64
        - name: positionId
          in: path
          description: ID of the position to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Branch position successfully deleted
        '404':
          description: 'Branch, department, or position not found'
        '500':
          description: Internal server error
...
  '/api/v1/branches/{branchId}/departments/{departmentId}/positions/filter':
    post:
      tags:
        - Branch Position Management
      summary: Get all positions for a branch department with filtering
      description: >-
        Returns a paginated list of positions for a specific branch department
        based on filter criteria
      operationId: filterBranchPositions
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: integer
            format: int64
        - name: departmentId
          in: path
          description: ID of the department
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBranchPositionDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved branch positions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchPositionDTO'
        '404':
          description: Branch or department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchPositionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchPositionDTO'
  '/api/v1/branches/{branchId}/departments/{departmentId}/positions':
    post:
      tags:
        - Branch Position Management
      summary: Create a new position for a branch department
      description: >-
        Creates a new position for a specific branch department with the
        provided details
      operationId: createBranchPosition
      parameters:
        - name: branchId
          in: path
          description: ID of the branch
          required: true
          schema:
            type: integer
            format: int64
        - name: departmentId
          in: path
          description: ID of the department
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchPositionDTO'
        required: true
      responses:
        '201':
          description: Branch position successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '400':
          description: Invalid branch position data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '404':
          description: Branch or department not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchPositionDTO'
...
components:
  schemas:
...
    BranchPositionDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        departmentId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    FilterRequestBranchPositionDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BranchPositionDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for branch positions
    PaginationResponseBranchPositionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BranchPositionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...